Refactor FiltersSidebar into modular components
